requirements.txt
discord.py
jikanpy